requirements.txt
alembic==1.0.11
cachelib==0.1
certifi==2019.6.16
chardet==3.0.4
Click==7.0
Flask==1.0.3
Flask-Migrate==2.5.2
Flask-SQLAlchemy==2.4.0
future==0.17.1
gunicorn==19.9.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
line-bot-sdk==1.11.0
Mako==1.0.13
MarkupSafe==1.1.1
psycopg2-binary==2.8.3
python-dateutil==2.8.0
python-editor==1.0.4
requests==2.22.0
six==1.12.0
SQLAlchemy==1.3.5
SQLAlchemy-Utils==0.34.0
urllib3==1.25.3
Werkzeug==0.15.4


